day20pt2: done
Like day8pt2, this is another solution that doesn't solve the general
case and makes assumptions about the nature of the input.

The lead-in to the final module takes inputs from four separate networks
that run in separate cycles, so this solution just drives each broadcast
output individually to detect those sub-network cycles, and prints the
cycles out so they can be plugged into an LCM calculator of your choice.

Light editing of the puzzle input allows it to be fed into graphviz,
which reveals the four sub-networks.