day25: implement solution in Rust
At first glance, the puzzle seems to be asking for the min cut that
divides the two sub-graphs, implying the use of an algorithm like
Stoer-Wagner or Karger, but since we already know that the min cut is
three edges, we can instead find the three shortest paths between two
random points, remove them all, and if there isn't a fourth, the
sub-graphs are divided.  From there, a simple flood fill from a single
vertex is enough to calculate the answer.